requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Flask==3.0.0
fonttools==4.43.1
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.6.3
mysql-connector-python==8.2.0
numpy==1.26.0
packaging==23.2
pandas==2.0.0
Pillow==10.0.1
protobuf==4.21.12
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
zipp==3.17.0
